// src/pages/area-comum/tabelas/entidades/centros-saude/types/centro-saude-edit-form-types.ts
export interface CentroSaudeEditFormValues {
  nome: string
  codigoLocalCS: string
  email: string
  numeroContribuinte: string
  observacoes: string
  status: number
  // Morada (Rua)
  paisId: string
  distritoId: string
  concelhoId: string
  freguesiaId: string
  codigoPostalId: string
  rua: string
  ruaId: string
  numeroPorta: string
  andarRua: string
}